Texture Garbage
Well there's a texture showing up... Textures sorta work but Josh needs
to figure out how to actually implement the pile of code added today.
Also fixed the name of IRenderObject to fit standards. Any other changes
are probably as a result of messing around. Credit to Josh for the fix
on alpha channels of textures (TexturedObject ln 45-46).