Stuff Was Done
Reorganized the Interfaces and Logic classes; replaced InteractionLogic
with CollisionLogic (to be implemented in the future) and IInteractable.
Cleaned out TexturedObject and added MovableObject to include the
movement code. Also moved texture registry into the WorldData class to
be loaded before the objects are added to sceen (safeguards will need to
be added later). Also set up the Input class to register key input and
map to the inputs in a dictionary, and send the key codes as enums to an
event that others can hook in to. Implemented this in the MovableObject
with WASD controls to move the heart around the scene.