Started Attack and Effect System
Added a few classes and interfaces for Effects and Attacks, with support
for lambda injection to Action objects that are called when needed. This
allows for a lot of flexibility when designing attacks and effects in
the future. Also reworked some of the logic to be thread safe now, and
implemented a standard mob to be both Effectable and Attackable. Also
goofed around with janky collision adjustment code, though it will be
necessary to include velocity in calculations for better adjustments.