UI System, Better Movement, Better Input
Added Button classes and the necessary methods and math to convert mouse
coordinates into the gl coordinates for click detection. Also set up a
better input system that allows for multiple keys to be held at the same
time, and smoother recognition. To top it off, the whole world movement
system was reworked to avoid having to adjust every world object. Now
just the orthographic GL projection is moved when the controllable
object moves, dramatically simplifying the math on the cpu. The screen
objects are rendered afterwards in a separate projection that doesn't
move with the character to keep static motion on the screen without the
aforementioned issues with calculations on the GUI objects. Also enabled
multisampling again, as it smooths out motion and texture edges at
smaller sizes. Also toyed with animation of objects using the texture
coordinates to change the sprite. Currently the texture is just
scrolling to the left, but it's a good demonstration of how to take a
whole texture and slice it up using just texture coords instead of
loading multiple textures from one spritesheet.