Worked on World Loading, Changed Ortho
This time around, the correct logic for loading and unloading based on
past regions is used, though it could be smoothed out to avoid lag when
crossing region boundaries. Also changed the orthographic projections to
integer sizes based on the screen width and height in pixel numbers,
though this hasn't aleviated much of pixel issues that exist; might
revert back to the previous method in the future. As a result of the
orthographic changes, world coordinates are messed up again, and as
such, world loading isn't quite right either, though that will be fixed
once the new Ortho design is settled. Also adjusted the default window
coordinates to be a perfect 16:9 ratio at a multiple of 62 (twice the
desired pixel number).