Adjusted Coordinates
Moved the Orthographic projection again into the 496x279 coordinate
system that was previously decided upon. Most things are updated to
reflect the move from float coordinates into quantized (hopefully)
texture units that are then mapped to the screen via the ortho. Also
fixed the region detection code, though it needs to be tweaked with the
coordinate system, as 100x100 is smaller than just what the screen
displays, 1000x1000 is most likely a temporary fix, though it might
stay. Also enabled mipmapping and messed around with the filtering
options: linear needs to be enabled for text, but keep nearest for the
rest of the textures.