New Texture Management System
Moved more data into the TextureManager class: the paths to textures and
their names are declared in the manager class, and each region or screen
file that needs textures will pass an array of the names they need to
the manager to be loaded if necessary. The textures will only then be
unloaded once all regions that needed it are unloaded. Also added thread
locking to the worldData objects and moved to HashSets instead of Lists,
as they are faster for finding specific objects (not used yet, but will
likely be used in the future) than Lists. Also fixed the spaghetti-code
in the WorldObjectAnimated class and got the MCObject working and
displayed once again, though animation isn't set up correctly in the
constructor yet.