Added Windows, Fixed Input
Added a windows system similar to the world region system, where each
window supplies its own list of needed textures and then a hashset of
the objects to be rendered. Screen now manages these windows instead of
rendering directly. Added a pause menu and the HUD to test
functionality, though more work would be needed to get the hud disabled
when other windows are open. Also improved input by allowing only
certain keys to be held down continuously; all other keys are only
triggered once per depression. Also added a GameState manager to handle
the state of the game that other classes can reference - it's not
teribly useful currently, but in the future it can handle loading,
menus, and other unique aspects of the game.