Solved Input Problems
Found a solution to the input problems: added an event for when the form
loses focus to deactivate all held keys, and added that same
functionality to resize events. It's still impossible to capture key
presses when resizing the window, but at least movement won't be stuck
(it will just be cancelled).