Improvements Regarding Screen Text Scaling
Text was looking wonky, so in order to fix the uneven mapping of text
texels onto display pixels, the screen projection was adjusted to a 360p
size by pixel count, not float values. All screen code was moved over to
a relative pixel system. 360p was used because it produces even
multiples on all standard display resolutions (720p, 1080p, 1440p, etc),
so as long as the pixels are lined up in the code on 360p, they will
always display correctly on any similar resolution. UIText and similar
objects were modified to use a textSize float that multiplies by 16 to
get a multiple of 16 pixels wide for all character textures and similar
code was updated to use the new size system. Currently, it's unclear
whether the world system will move to a similar system, though there is
little reason not to assuming other testing with this coordinate system
goes smoothly. Also adjusted the default window size to 720p.