Improved Dropdown, Started Save Files, World Editing
Dropdowns now only show a few elements at a time and are able to be
scrolled through. Small tweaks were made to improve loading and layering
of UI elements. Added Json.Net to handle serialization of data to file
and added classes to store data. Briefly outlined a UI window to hold
the world editor functions, though that functionality needs to be put on
hold until after world saves are complete, and runtime additions to
world data are possible.