World Saving is Working
Revamped the save system such that each world object is now responsible
for packing up and unpacking all data it needs to instantiate itself
upon next launch. The data is stored in the region's save file and
handled by such. The only shortcomings of this is that Actions
associated with interactable objects or buttons are unable to be parsed
and saved, though that kind of thing shouldn't really be stored anyway,
and some combination of reflection in the future could work around that
limitation. The save files are put in the build directory (so aren't
synced via Git) and can be generated via the console command 'saveGame'.