Added and implemented Character sprite
Added a character finally that looks useable. A bug is that the texture
is squished but I would be sitting here for a few hours just figuring
out what the hell is wrong. So I will just wait until my partner wakes
up in order to figure out how to squish. This sprite's dimensions sit at
37 : 22 pixels so I didnt expect our current system to be able to handle
it simply. Due to the fact that so far all sprites are 16 : 16 pixels.
Also implemented the character sprite by loading the texture and
painting it on the screen although it is a bit squished.

Note : THIS IS NOT THE MAIN CHARACTER. Its just for testing but might
have some placement later.