Quickly Added a character walk animation
Added a character walk animation so that my partner can maybe do some
testing with having a main character have a specific animation cycle for
a specific function instead of always having the character do it like
the flower. Went to bed and then woke up and decided I needed to do this
real quick.