'Finished' World Editor
The builder window is now working and modifications to variables will be
stored and saved. To use, select an object, and then scroll through the
variables list that appears, modifications are made upon enter or
clicking to another object. Remember to save (either by command or via
button at the bottom) to keep changes. Also added 2 more utility
textures and fixed the BaseFont 'N' again.