Builder Improvements, GUI Tweaks
World Builder can now add objects and remove objects from the world,
adding is quite complex, but the end result is that it pulls up the
Complete Declaration constructor for each object and will populate the
list as necessary for each variable. Not all errors are sorted out yet,
(there are a lot of errors involved with strings into variables into new
objects), but currently it just won't do anything if it runs into an
error along the way. Also added more support for hiding UI elements,
including hiding the to button of a dropdown to make just a scroll
panel.