Developed Storage of Interactable Objects
TLDR: Lots and lots of recursion. Also removed the ref passing of
LoadObjects() and GetObjects() in windows to improve .NET version
compatability (and it wasn't really needed). Scroll bar is pretty much
finished, it supports scrolling as well as clicking and dragging, and is
linked to scrollable interfaces both ways now. Also improved readability
of save files with regards to enums, and renamed RegionSave to GameSave
to reflect the added functionality.