New Editor Window
After coding the entire system for adding Interaction Elements to the
custom editor window, it became apparent that the custom window solution
was not going to work. A new form-based design was created as DevWindow,
to hold all development tools needed for the game. For interaction, a
display tree was used to show the dialogue succinctly, with
modification, addition, and deletion all implemented. The main overview
tab wasn't touched much other than to add a demo system to update and
show world coordinates. The only thing missing from the interaction
editor is renaming choice paths, and moving orders around. And of course
everything involving Quest data.