Fixed Text Spacing
A ridiculous problem solved with a ridiculous solution. This will be the
first of many ridiculous solutions that resolve profoundly ridiculous
problems. What has happened hear today is not a triumph it tis just a
tired man at 3:15 in the morning looking like a total asshole because he
had to draw a diagram using the word fuck to try and figure out the
reason the text was "randomly shifted". In the end all was discovered as
it was relized that the pivet point for all the objects is located
directly in the middle of each object and not in the bottom left corner
as the previous code would have supported. This instead was a grand
problem as it cause some overlap or even extra spaces due to the fact
that most of the characters have a profound dissonance when it comes to
the characters width. This meant that if a character with a greater
width went to a character of lesser width it would cause extra spaces or
overlap based on the situation and how dramatic the difference was. It
is very similar to the situation when the character width goes from
lesser to greater. Let it be known to all who may read this github
commit. The solution was found. Woot.

Oh I also changed the n, and p characters to look a bit nicer as well as
the copy pasta was changed into something a bit more usefull. The copy
pasta will be implemented somewhere else but currently its less
important.