Implemented InteractionQuest in the Dev Window
Everything with quest data, immediate mode and quest tags are now
visible and editable in the interaction tree editor, and quests are sort
of working in interaction windows, though they can't remove based on
quest flag yet; that will get added once quest parsing is set up. Also
fixed hiding and unintentional setting of width and height in world
bases, and added saving of width and height for all objects.