Bug Fixes, Beginning Quest Work
Fixed issues including interaction editors, world builders, and
notifications. Also adjusted the texturePosition float arrays to be
static if possible to reduce memory footprint in the future (not much
difference currently, but will help with static objects in the future.)
Quest work started today with the development of simple item collection
tasks outlined and serialized to storage. More work will be done in the
future and functionality expanded as needed (and possible). The system
currently only supports single-dialogue interactions, and the link
between quest data and interaction/gui isn't set up yet.