Interaction Cleanup
Smoothed out the Interaction Window, added key and mouse functionality
instead of back and next buttons, and shifted the window down to a more
appropriate location. WorldObjectInteractable has been simplified to
solely a dialogue opening object, so more world objects and a rename
will be needed in the future to facilitate inventories and other guis.