Added Key Bindings to Settings
Set up a tab system in the options window and added a tab for controls.
The current key binding system seems to be quite flexible, and should
allow all possible inputs from the keyboard other than the reserved keys
for escape and console. Also made slight adjustments to the dev window.
Fixed bugs with the console window, and added better flexibility to the
Input system.