Save Changes
Added data packages to store game region and quest data (numbered in
that order.) The Package.rzz files are zip archives that are read and
written to with a new dll, DotNetZip. Other methods and saving systems
have been update to reflect the changes, including a new GameData
directory for game files, leaving the SaveData directory for player
settings.