Basic Pathfinding AI
Started work on the AI system, wrote a pathfinding system based on A*
(though not complete, as it can't pathfind around walls yet). Adjusted
the world system a bit, and made room for Node hashes in the world file.
Node and NodeGrid are for pathfinding. Made improvements as necessary to
StandardMob and Movable world objects. Changed the sprite loader to use
the development location instead of the output directory, will move to
archive reading at some point in the future (will need an automated
build option to package the sprites directory into a zip on compile) and
restored Character.png for use as a mob texture.