Added Item System
Created an item system for the game. A few interfaces and base classes
were created to begin the item system. Utils added were the Item
Dictionary for new creation of item objects, and console commands to
create items from command line. Also added the inventory window with an
item list view, and added clicking functionality for equipping and
consuming. Serialization and Deserialization from disk was done via a
similar method to game objects, though the enum system was ignored for
simplicity's sake (might move the game object system over to type based
storage instead of enum in the future. )