UI Improvements
Fixed Setup and Cleanup on elements in tabs - fixing dropdown bugs - and
fixed bugs with input getting hung on controls changes in the options
menu. Added a selection outline and functionality to the item grid
(though methods for using the selected object are not implemented yet)
and added history functionality for the console via up/down arrow keys.