Other Inventories, Small Features
Added worldobject inventories, a chest-esque and an item-stack-esque
object, along with all the serialization structs and methods for
storage. Also made some tweaks to console and added a tab to the
devWindow to handle nearby inventories and item modification. The
standard for worldobject layers has been moved from 5 to 10 to make more
room for GUI layers; keep this in mind in the future.