Happy New Years!  -Multiplayer support, Reorganization
WorldObjects got a GUID for use in identifying each object in the world,
and WorldSave packaging was reorganized to better use inheritance.
Resorted the code for and save locations: there's now support for
multiple save folders, each folder containing all player and world data,
and a master folder that is copied to each new save directory,
Package_03 now stores the player data files. Bugs with player movement
being jittery are solved with new ortho update events to synchronize
movement with the orthographic projection changes on each render loop.
BuilderWindow had several bugs fixed.

Networking was added to allow cooperative play between a host game and a
guest game. The Net package contains the UDPFrame packet that's
instantiated and serialized via the ProtoBuf standard, the UDPMessager
class to handle all of the networking retrieval and sending, and the
SessionManager to handle synchronization and management between the
connected clients. UDPFrame can handle arrays of any length for all data
types necessary, and has a packetID that's assigned upon submission for
tracking to allow lossless packets (using confirmation returns and
resending to ensure data is sent). Current networking supports
connection between two games, one hosting, one joining, and synchronized
player movement. Movement of other entities and objects in the game is
added, but not working perfectly as of this commit. The transmission of
WorldObject data is roughed out, with ToPacket storing all necessary
data in a UDPFrame for transmission, and PacketToWorldObject in
WorldData to update the WorldObject information on the other side, but
neither is correctly implemented, and a shift of focus as to how to deal
with world data is required before continued development. All networking
features are available through somewhat extensive console commands.
Player data storage and serialization was also reworked to support
multiple player files, and correct storage of all data necessary for the
player; PlayerData will, in the future, hold more information, like
current quests, stats, etc. Apologies in advance for extra WriteLines
that haven't been removed from testing yet.