New Texturing System, New Save System
Textures were moved to a completely new system that should be faster,
more convenient for development, and uses less memory. The entire save
system was shifted to Binary serialization from JSON serialization (with
the exception being the settings file). All systems that used packaged
save objects to serialize the data and construct are gone, with direct
binary serialization taking a much simpler approach. Regions were
reworked as well, with custom RegionTag objects instead of string tags
to improve efficiency in finding and managing regions. Entities are not
currently supported, as there's major fixing to be done with the entity
inheritance structure and storage. Most networking code concerning world
objects is now invalid and (hopefully) commented out. Inventory, Quest,
and Interaction storage is set up but not tested yet. Also fixed #7.