Entity Improvements, Texture Loading Fixes
Started work on Entity codebase. Entities are now stored in a hashset
for the whole world, loaded from a file in Package_01. Added an
IDelayable interface to allow any object to count animation ticks, and
added the framework for decaying entities (e.g. particles) to be added
later. Moved some mob code around and renamed the inventory class to
make room for future entities that have an inventory. As it turns out,
the new texture system had issues with serializing static object
references - it would just construct new objects upon deserialization.
The new solution uses a dictionary lookup table as a list of all
textures added to the TextureNames class and finds the reference once
constructed.